+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchModel.kt
@@ -1,13 +1,18 @@
 package org.simple.clinic.home.overdue.search
 
 import android.os.Parcelable
+import androidx.paging.PagingData
+import kotlinx.parcelize.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
+import org.simple.clinic.home.overdue.OverdueAppointment
 
 @Parcelize
 data class OverdueSearchModel(
     val overdueSearchHistory: Set<String>?,
     val searchQuery: String?,
-    val overdueSearchProgressState: OverdueSearchProgressState?
+    val overdueSearchProgressState: OverdueSearchProgressState?,
+    @IgnoredOnParcel
+    val overdueSearchResults: PagingData<OverdueAppointment> = PagingData.empty()
 ) : Parcelable {
 
   val hasSearchQuery: Boolean
@@ -36,5 +41,9 @@ data class OverdueSearchModel(
 
   fun loadStateChanged(overdueSearchProgressState: OverdueSearchProgressState): OverdueSearchModel {
     return copy(overdueSearchProgressState = overdueSearchProgressState)
+  }
+
+  fun overdueSearchResultsLoaded(overdueSearchResults: PagingData<OverdueAppointment>): OverdueSearchModel {
+    return copy(overdueSearchResults = overdueSearchResults)
   }
 }