+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchUiRenderer.kt
@@ -30,6 +30,8 @@ class OverdueSearchUiRenderer(
         // No-op
       }
     }
+
+    ui.setOverdueSearchResultsPagingData(model.overdueSearchResults, model.searchQuery.orEmpty())
   }
 
   private fun renderResults() {