+++ app/src/main/java/org/simple/clinic/util/PagerFactory.kt
@@ -4,8 +4,10 @@ import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import androidx.paging.PagingSource
+import androidx.paging.rxjava2.cachedIn
 import androidx.paging.rxjava2.observable
 import io.reactivex.Observable
+import kotlinx.coroutines.CoroutineScope
 import javax.inject.Inject
 
 typealias PagingSourceFactory<K, V> = () -> PagingSource<K, V>
@@ -26,5 +28,22 @@ class PagerFactory @Inject constructor() {
         initialKey = initialKey,
         pagingSourceFactory = sourceFactory
     ).observable
+  }
+
+  fun <K : Any, V : Any> createPager(
+      sourceFactory: PagingSourceFactory<K, V>,
+      pageSize: Int,
+      enablePlaceholders: Boolean,
+      initialKey: K? = null,
+      cacheScope: CoroutineScope
+  ): Observable<PagingData<V>> {
+    return Pager(
+        config = PagingConfig(
+            pageSize = pageSize,
+            enablePlaceholders = enablePlaceholders
+        ),
+        initialKey = initialKey,
+        pagingSourceFactory = sourceFactory
+    ).observable.cachedIn(cacheScope)
   }
 }